% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreDE.R
\name{exploreDE}
\alias{exploreDE}
\title{start exploreDE application}
\usage{
exploreDE(dir = NA)
}
\description{
start exploreDE application
}
\examples{
\dontrun{
myDir <- "/srv/GT/analysis/peter/rs_connect_apps2/exploreDE/SummarizedExperiment.rds"
exploreDE(myDir)
}
}
