% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{generateParams}
\alias{generateParams}
\title{generateParams}
\usage{
generateParams(
  sampleGroup = "diff",
  refGroup = "undiff",
  groupingName = "Condition",
  comparison = "diff--over--undiff",
  featureLevel = "gene",
  runGO = TRUE,
  fdrThreshORA = 0.05,
  fdrThreshGSEA = 0.05,
  pValThreshGO = 0.01,
  log2RatioThreshGO = 0,
  backgroundExpression = 10,
  refBuild = "GRCh38"
)
}
\description{
generateParams
}
