% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreDE.R
\name{exploreDE}
\alias{exploreDE}
\title{start exploreDE application}
\usage{
exploreDE(fileSE = NA, demo = NA)
}
\description{
start exploreDE application
}
\examples{
\dontrun{
myDir <- system.file("examples/proteomics/SummarizedExperiment.rds", package = "exploreDE")
exploreDE(myDir)
}
}
